Move menu options into a folder 'MenuOptions'
